requirements.txt
pandas
matplotlib
seaborn
scipy
streamlit
plotly.express
